+++ dfetch/__init__.py
@@ -1,5 +1,5 @@
 """Dfetch."""
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 
 DEFAULT_MANIFEST_NAME: str = "dfetch.yaml"